more refactor on fiber-types
